refactor: rename proxied.ts → methods.ts
